feat: entity 설계
