+++ src/main/kotlin/io/hhplus/architecture/domain/lecture/model/Lecture.kt
@@ -1,23 +1,16 @@
-package io.hhplus.architecture.domain.lecture
+package io.hhplus.architecture.domain.lecture.model
 
-import io.hhplus.architecture.domain.model.common.BaseModel
-import io.hhplus.architecture.domain.user.User
-import jakarta.persistence.*
+import io.hhplus.architecture.domain.common.model.BaseModel
+import jakarta.persistence.Entity
+import jakarta.persistence.GeneratedValue
+import jakarta.persistence.GenerationType
+import jakarta.persistence.Id
 
 @Entity
 class Lecture(
     var title: String,
 
     var description: String,
-
-    var capacity: Int,
-
-    @OneToOne
-    @JoinColumn(name = "instructor_id")
-    var instructor: User,
-
-    @OneToMany(mappedBy = "lecture")
-    val scheduleList: MutableList<Schedule> = mutableListOf()
 ) : BaseModel() {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
