--- src/main/kotlin/io/hhplus/architecture/domain/lecture/model/LectureStatus.kt
@@ -1,5 +1,0 @@
-package io.hhplus.architecture.domain.lecture
-
-enum class LectureStatus {
-    CANCELED, PENDING, CONFIRMED
-}
